requirements.txt
Flask
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
requests
msal
